uCoreMinus: initial deeply embedded version
It's not fully correct: as we added dependent types, the type itself
should be defined in context. For now I solved by adding universal
quantification over the context to the dependent type constructors but
that's not correct: that essentially packs arbitrary context together
with the term well-typed in this context into the constructor. Whereas
in reality we want the term inside the constructor to be well-typed
w.r.t. current context, not some arbitrary one.

But making Ty depend on context makes the definition of context as just
a list of Ty's impossible, so this problem is not so easy to
tackle. I'll try to handle it in future.