uCoreMinus: seemingly more correct version
Version with types defined in context. Context itself is no longer just
a list of types. Now each new type is defined in the context it's added
to. That's a questionable representation (as I had to add the weaken
rule to Ty) but I haven't managed to find another one.

Hopefully, it will be possible to define a denotational semantics for
this version.